Use mjs for rollup config
